requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.5
ci-info==0.2.0
click==8.0.1
configobj==5.0.6
configparser==5.0.2
distro==1.6.0
Django==3.2.7
django-cleanup==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
ftfy==6.0.3
future==0.18.2
httplib2==0.19.1
idna==3.2
isodate==0.6.0
lxml==4.6.3
networkx==2.6.3
nibabel==3.2.1
nipype==1.6.1
numpy==1.21.2
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.3
pathlib==1.0.1
pdftotext==2.2.0
Pillow==8.3.2
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.18.17
pyparsing==2.4.7
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
pyxnat==1.4
rdflib==6.0.0
regex==2021.8.28
requests==2.26.0
scipy==1.7.1
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
tabula-py==2.3.0
traits==6.2.0
urllib3==1.26.6
wcwidth==0.2.5
